requirements.txt
streamlit>=1.37.0
faiss-cpu

huggingface-hub>=0.23.0

langchain>=0.3.0
langchain-openai>=0.1.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0

openai>=1.0.0
numpy
python-dotenv
sentence-transformers
transformers
torch

# For Unsloth fine-tuned models (comment out if you don't use it)
unsloth

# Multi-agent orchestration
agno>=0.4.0

ragas>=0.1.10
datasets>=2.14.0

# API server
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

